Numeric.ModalInterval

Description

Kaucher Interval arithmetic based on intervals library by (c) Edward Kmett 2010-2013 and https://sites.google.com/site/modalintervalcalculator/

Documentation

data Kaucher a

Kaucher arithmetic can describe a proper (forall) and improper (exists) intervals with different semantics.

The interval `K x y` is proper if `x >= y` and improper otherwise.

(+/-) :: RealFloat a => a -> a -> Kaucher a infixl 6

(<.<) :: RealFloat a => a -> a -> Kaucher a infix 3

Create open and closed proper intervals

>>> 1 <=.<= 3
∀ x . x ∈ [1, 3]

>>> 3 <=.<= 1
∃ x . x ∈ [1, 3]

proper :: Ord a => Kaucher a -> Kaucher a

improper :: Ord a => Kaucher a -> Kaucher a

whole :: RealFloat a => Kaucher a

empty :: RealFloat a => Kaucher a

isEmpty :: Kaucher a -> Bool

singleton :: RealFloat a => a -> Kaucher a

inf :: Kaucher a -> Maybe a

sup :: Kaucher a -> Maybe a

dual :: Kaucher a -> Kaucher a

width :: Num a => Kaucher a -> a

isInvalid :: Kaucher a -> Bool

bisect :: Fractional a => Kaucher a -> (Kaucher a, Kaucher a)

Bisect an interval at its midpoint.

>>> bisect (10.0 ... 20.0)
(10.0 ... 15.0,15.0 ... 20.0)

>>> bisect (singleton 5.0)
(5.0 ... 5.0,5.0 ... 5.0)

>>> bisect Empty
(Empty,Empty)

midpoint :: Fractional a => Kaucher a -> a

Nearest point to the midpoint of the interval.

>>> midpoint (10.0 ... 20.0)
15.0

>>> midpoint (singleton 5.0)
5.0

>>> midpoint empty
*** Exception: empty interval

member :: Ord a => a -> Kaucher a -> Bool

Determine if a point is in the interval.

>>> member 3.2 (1.0 ... 5.0)
True

>>> member 5 (1.0 ... 5.0)
True

>>> member 1 (1.0 ... 5.0)
True

>>> member 8 (1.0 ... 5.0)
False

>>> member 5 empty
False

notMember :: Ord a => a -> Kaucher a -> Bool

Determine if a point is not included in the interval

>>> notMember 8 (1.0 ... 5.0)
True

>>> notMember 1.4 (1.0 ... 5.0)
False

And of course, nothing is a member of the empty interval.

>>> notMember 5 empty
True

(|^|) :: (RealFloat a, Ord a, Integral b) => Kaucher a -> b -> Kaucher a infixr 8

(|^^|) :: (RealFloat a, Ord a, Integral b) => Kaucher a -> b -> Kaucher a infixr 8

intersection :: (RealFloat a, Ord a) => Kaucher a -> Kaucher a -> Kaucher a

Calculate the intersection of two intervals.

>>> intersection (1 ... 10 :: Interval Double) (5 ... 15 :: Interval Double)
5.0 ... 10.0

hull :: Ord a => Kaucher a -> Kaucher a -> Kaucher a

Calculate the convex hull of two intervals

>>> hull (0 ... 10 :: Interval Double) (5 ... 15 :: Interval Double)
0.0 ... 15.0

>>> hull (15 ... 85 :: Interval Double) (0 ... 10 :: Interval Double)
0.0 ... 85.0

meet :: (RealFloat a, Ord a) => [Kaucher a] -> Kaucher a

join :: (RealFloat a, Ord a) => [Kaucher a] -> Kaucher a

contains :: Ord a => Kaucher a -> Kaucher a -> Bool

Check if interval X totally contains interval Y

>>> (20 ... 40 :: Interval Double) `contains` (25 ... 35 :: Interval Double)
True

>>> (20 ... 40 :: Interval Double) `contains` (15 ... 35 :: Interval Double)
False

isSubsetOf :: Ord a => Kaucher a -> Kaucher a -> Bool

Flipped version of contains. Check if interval X a subset of interval Y

>>> (25 ... 35 :: Interval Double) `isSubsetOf` (20 ... 40 :: Interval Double)
True

>>> (20 ... 40 :: Interval Double) `isSubsetOf` (15 ... 35 :: Interval Double)
False