Numeric.ModalInterval.Algorithms

Description

Kaucher Interval arithmetic algorithms for outer and inner approximations based on https://tel.archives-ouvertes.fr/tel-00331668/document

Documentation

outerApprox :: Tree -> Domains -> Kaucher Double

Outer approximation of an expression tree given the domains of the variables. To calculate the outer approximation we need to apply "dualizer" to every variable that is totally monotonic.

See page 61 of https://tel.archives-ouvertes.fr/tel-00331668/document (Eq. 3.80)

innerApprox :: Tree -> Domains -> Kaucher Double

Inner approximation of an expression tree given the domains of the variables. To calculate the outer approximation we need to apply "dualizer" to every variable that is totally monotonic and replace those non-monotonic with the midpoint of the domain.

See page 61 of https://tel.archives-ouvertes.fr/tel-00331668/document (Eq. 3.79)